requirements.txt
selenium
beautifulsoup4
python-dotenv
pandas
tqdm
langchain
langchain-openai
openai
pydantic 